+++ tests/cmake_import/project/main.cpp
@@ -1,4 +1,5 @@
 #include <databento/record.hpp>
+#include <iostream>
 
 int main(int argc, char** argv) {
   databento::RecordHeader header{};