+++ tests/src/mock_tcp_server.cpp
@@ -4,9 +4,10 @@
 #ifdef _WIN32
 #include <winsock2.h>
 #else
-#include <netinet/in.h>  // IPPROTO_TCP, ntohs, sockaddr_in, TCP_NODELAY
-#include <sys/socket.h>  // AF_INET, listen, recv, send, setsocketopt, socket
-#include <unistd.h>      // close, socklen_t
+#include <netinet/in.h>   // IPPROTO_TCP, ntohs, sockaddr_in
+#include <netinet/tcp.h>  // TCP_NODELAY
+#include <sys/socket.h>   // AF_INET, listen, recv, send, setsocketopt, socket
+#include <unistd.h>       // close, socklen_t
 #endif
 
 #include <thread>  // this_thread